requirements.txt
Package                    Version
-------------------------- ------------
aiofiles                   25.1.0
aiohappyeyeballs           2.6.1
aiohttp                    3.13.5
aiosignal                  1.4.0
annotated-doc              0.0.4
annotated-types            0.7.0
anyio                      4.13.0
asttokens                  3.0.1
attrs                      26.1.0
babel                      2.18.0
beautifulsoup4             4.12.3
black                      24.4.2
blinker                    1.9.0
certifi                    2026.4.22
cffi                       2.0.0
charset-normalizer         3.4.7
click                      8.3.3
colorama                   0.4.6
comm                       0.2.3
contourpy                  1.3.3
coverage                   7.13.5
cryptography               48.0.0
cssselect                  1.4.0
cycler                     0.12.1
dash                       4.1.0
debugpy                    1.8.20
decorator                  5.2.1
dnspython                  2.8.0
email-validator            2.3.0
executing                  2.2.1
fastapi                    0.111.0
fastapi-cli                0.0.24
fastjsonschema             2.21.2
feedfinder2                0.0.4
feedparser                 6.0.12
filelock                   3.29.0
flask                      3.1.3
fonttools                  4.62.1
frozenlist                 1.8.0
ghp-import                 2.1.0
google-api-core            2.30.3
google-auth                2.50.0
google-cloud-bigquery      3.41.0
google-cloud-core          2.5.1
google-crc32c              1.8.0
google-resumable-media     2.8.2
googleapis-common-protos   1.74.0
grpcio                     1.80.0
grpcio-status              1.80.0
h11                        0.16.0
httpcore                   1.0.9
httptools                  0.7.1
httpx                      0.27.0
idna                       3.13
importlib-metadata         9.0.0
iniconfig                  2.3.0
ipykernel                  7.2.0
ipython                    9.13.0
ipython-pygments-lexers    1.1.1
itsdangerous               2.2.0
jedi                       0.20.0
jieba3k                    0.35.1
jinja2                     3.1.6
joblib                     1.5.3
jsonschema                 4.26.0
jsonschema-specifications  2025.9.1
jupyter-client             8.8.0
jupyter-core               5.9.1
kiwisolver                 1.5.0
lxml                       6.1.0
lxml-html-clean            0.4.4
markdown                   3.10.2
markdown-it-py             4.0.0
markupsafe                 3.0.3
matplotlib                 3.9.0
matplotlib-inline          0.2.1
mdurl                      0.1.2
mergedeep                  1.3.4
mkdocs                     1.6.0
mkdocs-get-deps            0.2.2
mkdocs-material            9.5.27
mkdocs-material-extensions 1.3.1
multidict                  6.7.1
mypy-extensions            1.1.0
nbformat                   5.10.4
nest-asyncio               1.6.0
newspaper3k                0.2.8
nltk                       3.9.4
numpy                      1.26.4
orjson                     3.11.8
packaging                  26.2
paginate                   0.5.7
pandas                     2.2.2
parso                      0.8.7
pathspec                   1.1.1
pillow                     12.2.0
platformdirs               4.9.6
plotly                     5.22.0
pluggy                     1.6.0
prompt-toolkit             3.0.52
propcache                  0.4.1
proto-plus                 1.27.2
protobuf                   6.33.6
psutil                     7.2.2
pure-eval                  0.2.3
pyarrow                    15.0.2
pyasn1                     0.6.3
pyasn1-modules             0.4.2
pycparser                  3.0
pydantic                   2.13.3
pydantic-core              2.46.3
pygments                   2.20.0
pymdown-extensions         10.21.2
pyparsing                  3.3.2
pyrefly                    0.8.0
pytest                     8.2.2
pytest-asyncio             0.23.7
pytest-cov                 5.0.0
python-dateutil            2.9.0.post0
python-dotenv              1.0.1
python-multipart           0.0.27
pytz                       2026.1.post1
pyyaml                     6.0.3
pyyaml-env-tag             1.1
pyzmq                      27.1.0
referencing                0.37.0
regex                      2026.4.4
requests                   2.32.3
requests-file              3.0.1
retrying                   1.4.2
rich                       15.0.0
rich-toolkit               0.19.7
rpds-py                    0.30.0
ruff                       0.5.6
scikit-learn               1.5.0
scipy                      1.17.1
seaborn                    0.13.2
setuptools                 82.0.1
sgmllib3k                  1.0.0
shellingham                1.5.4
six                        1.17.0
sniffio                    1.3.1
soupsieve                  2.8.3
stack-data                 0.6.3
starlette                  0.37.2
tenacity                   9.1.4
threadpoolctl              3.6.0
tinysegmenter              0.3
tldextract                 5.3.1
tornado                    6.5.5
tqdm                       4.66.4
traitlets                  5.14.3
typer                      0.25.1
typing-extensions          4.15.0
typing-inspection          0.4.2
tzdata                     2026.2
ujson                      5.12.0
urllib3                    2.6.3
uvicorn                    0.30.1
watchdog                   6.0.0
watchfiles                 1.1.1
wcwidth                    0.7.0
websockets                 16.0
werkzeug                   3.1.8
yarl                       1.23.0
zipp                       3.23.1
